+++ hosts/darwin/template-darwin/default.nix
@@ -23,6 +23,7 @@ in
 {
   imports = [
     ../common-darwin.nix
+    ../../../modules/darwin/home-manager.nix
   ];
 
   networking.hostName = "template-darwin";